Inject dialog rendering logic
